+++ src/rdf.internal.ts
@@ -21,12 +21,14 @@
 
 // @rdfjs/dataset is a CommonJS module, so named imports break in Node.
 // Thus, import the default export, then obtain the `dataset` property from there.
-import rdfJsDatasetModule from "@rdfjs/dataset";
 import type * as RdfJs from "@rdfjs/types";
+import { Store } from "n3";
 import type { IriString } from "./interfaces";
 import type { XmlSchemaTypeIri } from "./datatypes";
 
-export const rdfJsDataset = rdfJsDatasetModule.dataset;
+export const rdfJsDataset: RdfJs.DatasetCoreFactory<RdfJs.Quad>["dataset"] = (
+  quads: RdfJs.Quad[],
+) => new Store(quads);
 
 export const localNodeSkolemPrefix =
   "https://inrupt.com/.well-known/sdk-local-node/" as const;