+++ CHANGELOG.md
@@ -16,6 +16,11 @@ The following changes are pending, and will be applied on the next major release
 
 The following changes have been implemented but not released yet:
 
+### Breaking Changes
+
+- Use the global `fetch` function instead of `@inrupt/universal-fetch`. This means this library now only works
+  with Node 18 and higher.
+
 ## [1.30.2] - 2023-09-26
 
 ### Bugfixes